New API names for main serializer class
Going from Serializer with #serialize() and #deserialize() to Marshal
with #dump() and #load().

There's no logical reason for the name change. Looking at a class whose
name includes "Serializer" makes me think that it shouldn't have a
deserialize(). At the same time, I want clients to use one driver class
when working with this library, so splitting out deserialize() and
serialize() into two classes wasn't an option.

The Marhsal#load() and #dump() is the naming used in ruby's Marshal
class. We'll try this naming for now.